% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{destinations_york}
\alias{destinations_york}
\title{Example destinations dataset: schools in York}
\description{
Example dataset from York, UK
}
\details{
See data-raw/zones_york.qmd for details on the data source.
}
\examples{
head(destinations_york)
}
\keyword{datasets}
