% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{od_aus}
\alias{od_aus}
\title{Example OD dataset: flows between regions in Australia}
\description{
Example dataset from Australia
}
\note{
Regenerate the data with scripts in the \code{data-raw} directory.
}
\examples{
head(od_aus)
}
\keyword{datasets}
