% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{od_census_min}
\alias{od_census_min}
\title{Example OD dataset}
\description{
Example OD dataset from the 2011 UK Census
}
\note{
Regenerate the data with scripts in the \code{data-raw} directory.
}
\keyword{datasets}
