% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si-package.R
\docType{package}
\name{si-package}
\alias{si}
\alias{si-package}
\title{si: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/robinlovelace/si}
  \item \url{https://robinlovelace.github.io/si/}
  \item Report bugs at \url{https://github.com/robinlovelace/si/issues}
}

}
\author{
\strong{Maintainer}: Robin Lovelace \email{rob00x@gmail.com} (\href{https://orcid.org/0000-0001-5679-6536}{ORCID})

Authors:
\itemize{
  \item Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})
}

}
\keyword{internal}
