% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si.R
\name{si_model}
\alias{si_model}
\title{Run a spatial interaction model}
\usage{
si_model(od, fun, var_p, ...)
}
\arguments{
\item{od}{A data frame representing origin-destination data, e.g. as created by
\code{\link[=si_to_od]{si_to_od()}}}

\item{fun}{A function that calculates the interaction (e.g. the number of trips)
between each OD pair}

\item{var_p}{Use this argument to run a 'production constrained' SIM.
Character string corresponding to column in the \code{od}
dataset that constrains the total 'interaction' (e.g. n. trips) for all OD pairs
such that the total for each zone of origin cannot go above this value.}

\item{...}{Arguments passed to \code{fun}}
}
\description{
Executes a spatial iteraction model based on an OD data frame
and user-specified function
}
\examples{
od = si_to_od(zones, zones)
fun_dd = function(od, d = "distance_euclidean", beta = 0.3) exp(-beta * od[[d]] / 1000)
od_dd = si_model(od, fun = fun_dd)
plot(od$distance_euclidean, od_dd$res)
fun = function(od, O, n, d, beta) od[[O]] * od[[n]] * exp(-beta * od[[d]]/1000)
od_output = si_model(od, fun = fun, beta = 0.3, O = "origin_all", n = "destination_all", d = "distance_euclidean")
head(od_output)
plot(od$distance_euclidean, od_output$res)
od_pconst = si_model(od, fun = fun, beta = 0.3, O = "origin_all", n = "destination_all", d = "distance_euclidean", var_p = origin_all)
plot(od_pconst$distance_euclidean, od_pconst$res_constrained)
plot(od_pconst["res_constrained"], logz = TRUE)
}
