% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{si_oa_wpz}
\alias{si_oa_wpz}
\title{Origin-Destination Data for Leeds}
\description{
This dataset contains origin-destination data for Leeds, including the number of trips between output areas (OAs) and workplace zones (WPZs).
}
\examples{
head(si_oa_wpz)
}
\keyword{datasets}
