% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{si_oa_wpz_d}
\alias{si_oa_wpz_d}
\title{Destination Data for Leeds}
\description{
This dataset contains the number of trips destined for each workplace zone (WPZ) in Leeds.
}
\examples{
head(si_oa_wpz_d)
sf:::plot.sf(si_oa_wpz_d["n_d"])
}
\keyword{datasets}
