% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{si_oa_wpz_o}
\alias{si_oa_wpz_o}
\title{Origin Data for Leeds}
\description{
This dataset contains the number of trips originating from each output area (OA) in Leeds.
}
\examples{
head(si_oa_wpz_o)
sf:::plot.sf(si_oa_wpz_o["n_o"])
}
\keyword{datasets}
