% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{si_od_census}
\alias{si_od_census}
\title{Example OD dataset}
\description{
Example OD dataset from the 2011 UK Census
}
\note{
Regenerate the data with scripts in the \code{data-raw} directory.
}
\examples{
head(si_od_census)
}
\keyword{datasets}
