% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si_model.R
\name{si_predict}
\alias{si_predict}
\title{Predict si model based on pre-trained model}
\usage{
si_predict(od, model, constraint_p, output_col = "interaction")
}
\arguments{
\item{od}{A data frame representing origin-destination data, e.g. as created by
\code{\link[=si_to_od]{si_to_od()}}}

\item{model}{A model object, e.g. from \code{\link[=lm]{lm()}} or \code{\link[=glm]{glm()}}}

\item{constraint_p}{Use this argument to run a 'production constrained' SIM.
Character string corresponding to column in the \code{od}
dataset that constrains the total 'interaction' (e.g. n. trips) for all OD pairs
such that the total for each zone of origin cannot go above this value.}

\item{output_col}{Character string containing the name of the new output
column. \code{"interaction"} by default.}
}
\description{
Predict si model based on pre-trained model
}
\examples{
od = si_to_od(si_zones, si_zones, max_dist = 4000)
m = lm(od$origin_all ~ od$origin_bicycle)
od_updated = si_predict(od, m)
}
\seealso{
si_calculate
}
