% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{si_pubs}
\alias{si_pubs}
\title{Example destinations dataset: pubs in Leeds}
\description{
Example dataset from Leeds, UK
}
\note{
Regenerate the data with scripts in the \code{data-raw} directory.
}
\examples{
head(si_pubs)
}
\keyword{datasets}
