% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si.R
\name{si_to_od}
\alias{si_to_od}
\title{Prepare OD}
\usage{
si_to_od(origins, destinations, max_euclidean_distance = 10000)
}
\arguments{
\item{origins}{\code{sf} object representing origin locations/zones}

\item{destinations}{\code{sf} object representing destination locations/zones}

\item{max_euclidean_distance}{Maximum distance in meters between OD pairs.
Only OD pairs that are this distance apart or less will be returned.}
}
\description{
Prepares an OD data frame that next could be used to
estimate movement between origins and destinations
with a spatial interaction model
}
\examples{
origins = centroids
destinations = centroids
odsf = si_to_od(origins, destinations)
plot(odsf)

}
