% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{si_zones}
\alias{si_zones}
\alias{si_centroids}
\title{Example zones and centroids}
\description{
\code{si_zones} and \code{si_centroids} represent administrative zones between which
flows are to be estimated.
}
\note{
The schema data can be (re-)generated using code in \code{data-raw}
}
\examples{
si_zones
sf:::plot.sfg(si_zones$geometry)
sf:::plot.sfg(si_centroids$geometry, add = TRUE)
}
\keyword{datasets}
