% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simodels-package.R
\docType{package}
\name{simodels-package}
\alias{simodels}
\alias{simodels-package}
\title{simodels: Flexible Framework for Developing Spatial Interaction Models}
\description{
Develop spatial interaction models (SIMs). SIMs predict the amount of interaction, for example number of trips per day, between geographic entities representing trip origins and destinations. Contains functions for creating origin-destination datasets from geographic input datasets and calculating movement between origin-destination pairs with constrained, production-constrained, and attraction-constrained models (Wilson 1979) \doi{10.1068/a030001}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/robinlovelace/simodels}
  \item \url{https://robinlovelace.github.io/simodels/}
  \item Report bugs at \url{https://github.com/robinlovelace/simodels/issues}
}

}
\author{
\strong{Maintainer}: Robin Lovelace \email{rob00x@gmail.com} (\href{https://orcid.org/0000-0001-5679-6536}{ORCID})

Authors:
\itemize{
  \item Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})
}

}
\keyword{internal}
