% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zones}
\alias{zones}
\alias{centroids}
\title{Example zones and centroids}
\description{
\code{zones} and \code{centroids} represent administrative zones between which
flows are to be estimated.
}
\note{
The schema data can be (re-)generated using code in \code{data-raw}
}
\keyword{datasets}
