% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zones_aus}
\alias{zones_aus}
\title{Example zones dataset: regions of Australia}
\description{
Example dataset from Australia
}
\note{
Regenerate the data with scripts in the \code{data-raw} directory.
}
\examples{
head(zones_aus)
}
\keyword{datasets}
