% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zones_york}
\alias{zones_york}
\title{Example zones dataset: administrative zones of York}
\description{
See data-raw/zones_york.qmd for details on the data source.
}
\examples{
head(zones_york)
sf:::plot.sfg(zones_york$geometry)
}
\keyword{datasets}
