% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{si_oa_wpz_d}
\alias{si_oa_wpz_d}
\title{Destination Data for Leeds}
\description{
This dataset contains the number of trips destined for each workplace zone (WPZ) in Leeds.
}
\details{
See \href{https://wicid.ukdataservice.ac.uk/flowdata/cider/wicid/downloads.php}{wicid.ukdataservice.ac.uk} for details on the data source
and the file \code{data-raw/si_oa_wpz.qmd} in the package repo for details on how the example dataset was generated.
}
\examples{
head(si_oa_wpz_d)
sf:::plot.sf(si_oa_wpz_d["n_d"])
}
\keyword{datasets}
