% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/si.R
\name{si_to_od}
\alias{si_to_od}
\title{Prepare OD}
\usage{
si_to_od(origins, destinations, max_dist = Inf)
}
\arguments{
\item{origins}{\code{sf} object representing origin locations/zones}

\item{destinations}{\code{sf} object representing destination locations/zones}

\item{max_dist}{Euclidean distance in meters (numeric).
Only OD pairs that are this distance apart or less will be returned
and therefore included in the SIM.}
}
\description{
Prepares an OD data frame that next could be used to
estimate movement between origins and destinations
with a spatial interaction model
}
\examples{
origins = centroids
destinations = centroids
odsf = si_to_od(origins, destinations)
plot(odsf)

}
